backend_hf/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
timm
grad-cam
requests
chromadb
sentence-transformers
python-dotenv
reportlab

# MedGemma / Critic Agent
transformers>=4.38.0
accelerate>=0.27.0
huggingface_hub>=0.20.0
gradio_client>=0.8.0
scikit-learn>=1.3.0

langchain>=1.2.10
langgraph>=1.0.8
langchain-google-genai>=4.2.0

ultralytics>=8.0.0
